story test added.

+++ src/main/java/me/eting/server/core/repository/IncognitoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by lifenjoy51 on 12/14/14.
  */
-public interface IncognitoRepository extends JpaRepository<Incognito, Long> {
+public interface IncognitoRepository extends JpaRepository<Incognito, Integer> {
 
 }
